cmd/snap-update-ns: compute next action to transition mount profile
This patch adds a function that given two sorted mount profiles (current
and desired) computes the next action that needs to be taken to
transition the current profile to the desired profile.

The function operates on a simple principle. Identical entires are
skipped. Entries that don't show up in the desired profile but exist in
the current profile are umounted. Entries that are in the desired
profile but are not in the current profile are mounted.

The logic ensures that we always unmount everything that should be
unmounted before proceeding with the first thing to be mounted.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>