Error when `snap list foo` is run and no snap is installed
When `snap list` is run on a system without any snaps it returns
a zero exit code and a friendly message. But `snap list foo` also
does this and that is inconsistent with the behaviour of
`snap list foo` if some snaps are installed but not "foo". This
branch makes `snap list foo` fail in the same way regardless if
snaps are instaled or not.

LP: #1664383