cmd/libsnap: add function for validating snap names
This patch adds a function that can be used to validate snap names. Due
to the differences between go and libc regular expression engine the
same expression cannot be used but care has been taken to make sure it
has the same meaning.

The only difference is that the syntax "(?:)" becomes just "()"

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>