Disable running the configure hook on classic for the core snap
We see a large number of errors that looks like:
 `error: cannot find installed snap "core" at revision unset`
on core refrehes on classic. The reason is not yet understood.

We have no report about this error from an actual Ubuntu Core
system (this might be because the installed base is smaller there
though).

So far we disabled the configure hook for core but we need it on
Ubuntu Core systems because it contains important features like
disabling the sshd service and more. To unblock the situation
this branch will not run the configure hook for core on classic
(it is not very useful there anyway).