cmd/snap-update-ns: handle reused mount entries
This patch corrects a bug found by Paweł Stołowski where the algorithm
would unmount and re-mount and entry needlessly. This can be observed in
on the following input:

 desired = A B
 current = B C

The previous naive approach needlessly unmounted B and mounted it again
soon thereafter. The correct way to do it is to skip the unmount (and
subsequent mount) if the current entry can be found further down the
desired list.

This is achieved by tracking extra state in each mount entry (a flag)
and using this flag to track mount entries present in the desired
profile that were not unmounted and should not be mounted.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>