disable configure hook for core on classic
In PR#2958 the configure hook got disabled for classic because we
got a large number of errors on classic of the form:
  `error: cannot find installed snap "core" at revision unset`

However because on refresh the code that runs will insert a configure
task for the core snap. And only after the task got already inserted
snapd will re-exec. So we need to add the "ignore-core-hook-configure"
hack a bit later.