interfaces: add unity8 plug permissions (#2787)
This builds upon the u8-base PR (#2786) to add some permissions for unity8 apps that plug into the interface.

It's mostly some dbus allowances for session services that the unity8 session will provide. I tried to be somewhat minimal: I tested that a confined ubuntu-calculator-app can launch in a unity8 session with this. I haven't done extensive testing beyond that. I imagine we'll add more to this over time.

(Some of this is Ted's work repackaged.)