cmd/snap-update-ns: don't reuse mounts if parent changes
This patch fixes an issue found during unit testing where snap-update-ns
would reuse a mount entry even though the parent mount changed. This can
be illustarated with a simple example. The current/desired profiles both
contain: [A, A/B] but the nature of A is different (while keeping the
same directory name). The new algorithm correctly unmounts A/B and A and
then mounts A and A/B.

In addition rename the "flag" to "reuse" so that it has clear purpose.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>