release: detect if we are in ForcedDevMode by inspecting the kernel (#2952)
So far we are detecting if we need to go into "forced-devmode" by
inspecting the os-release file. The downside of this is that we
do not catch a lot of Ubuntu derivates that work just fine but
that we do not have in our code that checks /etc/os-release.

The result is that snapd thinks it is on a devmode distro and
will not generate an apparmor profile. But snap-confine is build
with the Ubuntu defaults which means it will have apparmor enabled
and expects a generated profile. The result is that no snap works
on a derived distro.

The new code will check via /sys/kernel if we are
running on a kernel that has the right apparmor bits enabled. Only if
this is not the case we will switch to "forced-devmode". This
will enable Mint 18.1, Zorin 12 and any other 16.04 derivate that
uses the Ubuntu kernel to work correctly.