tests: prevent automatic transition before setting the initial state of the test (#3017)
We are getting random errors on tests/main/classic-ubuntu-core-transition while setting up the test initial state. The installation of the test-snapd-python-webserver test snap after installing ubuntu-core can succeed or fail depending on the status of the automatically triggered ubuntu-core -> transition, if it has already started we get an error cannot install "test-snapd-python-webserver": ubuntu-core to core transition in progress, no other changes allowed until this is done and if it hasn't started yet the setup is done correctly.

These changes prevent the automatic transition by setting the ubuntu-core-transition-last-retry-time field in state.json, so that we can reliably set the initial conditions of the test.