tests: copy updated files with cp -a to preserve mode bits
When snap-confine was renamed to snap-wrap we noticed that snap-wrap is
no longer setuid root. This seems to be caused by "cp" vs "cp -a" done
in the test setup script. In the past this worked because both source
and destination were +s and then cp didn't have to create a new file.

This patch also adds a check that ensures snap-wrap is +s and that
snap-confine is not in the core snap if it was removed from the package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>