tests: remove stale apt proxy leftover from cloud-init
It seems that cloud init "marries" the image to the apt-proxy available
at the time the qemu image was built. While not common we should allow
moving a qemu image from one machine (with a set of apt proxy settings)
to another.

This patch removes the hard-coded apt proxy. Users interested in using
proxies have a range of environment settings at their disposal,
including HTTP_PROXY and APT_PROXY.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>