interfaces: fix for access denied of opengl interface
This is a fix for access denied of opengl interface when it's used
with framebuffer interface in the same snap.

Fix for https://bugs.launchpad.net/snapd/+bug/167573

Note: It's a short term fix without using udev tags for opengl interface
at this moment, not to break app snaps which were already released with
using the current opengl interface which doesn't use udev tags.

Signed-off-by: Chunsang Jeong <chunsang.jeong@canonical.com>