cmd: discard the C implementation of snap-update-ns
As the amount of logic increases the cost of writing snap-update-ns in C
was just mounting so this patch is an experiment to write it all in Go
instead.

If successful it can also be used as a test to implement some of the
snap-confine's shared parts and later on snap-confine itself.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>