Increase default max runtime for configure hook to 5min
To be on the safe side for now increase the maximum runtime for the
configure hook to 5 minutes to not run into a situation where the
limit is too low. Once we have more data on real world time we can
gradually decrease this timeout.

Also allow overriding this via SNAPD_CONFIGURE_HOOK_TIMEOUT in the
tests.