configstate,hookstate: timeout the configure hook after 5 mins, report failures to the errtracker (#3099)
Only allow all configure hooks to run for a limited amount of time. The
limit is 5 minutes currently to be on the safe side on very slow systems.
We may gradually decrease this timeout.

Also add support to report hook failures to the error tracker and report
failures in the core snap configure hook.

Allow overriding this via SNAPD_CONFIGURE_HOOK_TIMEOUT in the
tests.