cmd/snap-confine: always pass xauth file through regardless where it is located
Normally XAUTHORITY points to a file in /run/user/<UID> or /home/<user> and both
are allowed in the x11 interface for a snap to use. However in a few cases like
in KDE on openSUSE the file XAUTHORITY points to is placed in /tmp which doesn't
work with our current setup. As /tmp is not shared with the host namespace we
need to take the content of the file XAUTHORITY points to through and restore
it inside the snap environment.

Fixes LP #1677513