overlord: increase prune test wait by x10
We are seeing many failures that involve the TestEnsureLoopPrune test.
This test is inherently racy as it uses a actual time to wait for
something to happen. In case the machine is loaded heavily and other
processes contend for resources it may not get scheduled in time for
this to happen.

While the fix is not perfect I hope to at least decrease the frequency
of the failures. We may investigate a more correct fix where timing
would no longer be a factor if this is deemed insufficient.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>