overlord/ifacestate: ignore "ubuntu-core" if "core" candidates exist.
This patch changes how the interface manager processes auto-connect
candidates. In the case when both old ubuntu-core and the new core snaps
are installed (e.g. during a core transition) we now ignore any
candidates from the ubuntu-core snap. This way there is again only one
candidate and it can be auto-connected.

Fixes: https://bugs.launchpad.net/snappy/+bug/1680097
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>