interfaces: deduplicate return value from Repository.Connected
Because of missing validation of plug/slot names we don't really have
the guarantee that a given snap has unique plug or slot names. That
is fixed separately but we need to limit the damage. The aforementioned
method would return duplicate results when a snap is self-connected
and the plug and slot name is the same, as can happen with the "core"
snap and the "network-bind" plug and slot. This patch removes such
duplicates.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>