overlord/ifacestate: connect "network-bind" to core on startup
This patch cures the damage caused by a bug in the ubuntu-core -> core
transition. During that transition the new core snap would not get auto
connected and any new plugs it may have (such as the network-bind plug)
would stay disconnected. To fix this, the interface manager is now
automatically connecting (as if it was specified in the state) the
"core:network-bind" plug to the slot with the same name (yes this
should be forbidden but that is another bug)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>