cmd/snap-update-ns: add actual implementation
This patch adds a non-dummy implementation of snap-update-ns. There
are still three pieces missing. There's no locking so concurrently
running snap-confine is not synchronized. The function that determines
if a mount change is needed is dummy and always returns true.
The mount changes are not really performed yet as the Perform function
is just a stub.

The stubs will be addressed with separate PRs (one is already up), I just
wanted to land the general idea of how the tool operates.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>