interfaces: use safer sanitize functions
This patch switches the interface repository to the new pair of
Sanitize{Plug,Slot} functions that recover panics if the interface
sanitization code panics.

The interface code is processing user input and like it was found in the
following forum thread, it is not immune to errors. This code will
prevent denial-of-service attacks where snapd would otherwise crash when
presented with a potentially malicious snap.

Forum: https://forum.snapcraft.io/t/snapd-service-doesnt-start/319/5
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>