interfaces/mount: add partial implementation of Change.Needed
This patch adds a partial implementation of Change.Needed that works for
things that aren't bind mounted. Bind mounting requires some additional
detection because the fstab entry doesn't contain the whole information
and that needs to be reconstructed by looking at mountinfo.

To be precise, the source of the bind mount is only defined indirectly.
In a simple bind mount /src -> /dst the /src is a "pointer" to whatever
is mounted there (cound be something mounted directly at /src or could
be a fragment of the bigger mount at /).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>