overlord/snapstate/backend,interfaces/mount: move ns management code.
This patch moves two functions for managing mount namespaces to the
mount interface. This is done because the update-ns call has to be done
from within the mount backend and the original location was picked
because the overlord makes one call to discard-ns and, at that time, the
mount backend was much more empty.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>