partition: use uenv.OpenWithFlags and OpenBestEffort
We saw some panics in on-hardware testing of raspberry pi 2 and 3 that
look like data corruption on the SD card. Currently the code would just
panic (uenv.Open panicks) but with the new updates it would return an
error instead.

Errors are better but they are still quite fatal since this is not a A/B
resource that we can fall back on. To combat this we can pass the
OpenBestEffort flag. In this case parsing errors are ignored and
malformed data is discarded but other data is retained. This should let
us fare better in cases like this.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>