cmd/snap: correct devmode note for anomalous state
Past versions of snapd had a bug where snapd would lose the various
flags (like devmode) as a result of disable/enable operations.
In such state a snap would behave as if it was installed in jailmode
(the devmode flag was *not* set and confinement was strict) but the
output of "snap list" would still show it as devmode.

The cause of this is that in absence of the devmode flag we would look
at confinement type (what the snap declares the confinement should be).
This discrepancy between how the snap operates (strict) and how we show
it (devmode) can cause unwanted confusion. While we want to show an
useful "corrupted state" message later on we should not complicate how
the devmode note is computed. This patch makes us derive the devmode
note directly from the devmode flag.

Forum: https://forum.snapcraft.io/t/anomalous-snap-list-devmode-flag/699
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>