interfaces/builtin: Allow the use of timedatectl
Systemd's timedatectl utility can be used to set the time, timezone, and
enable/disable time server usage. It is a wrapper for the D-Bus API that
is already allowed in the time-control, timezone-control, and
timeserver-control interfaces. Snaps should be allowed to use the
utility in addition to using the D-Bus API.

Fix for https://forum.snapcraft.io/t/managing-time-date-and-timezone-in-ubuntu-core/408/5

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>