many: derive implicit slots from interface meta-data
This patch moves the declaration that a given interface should have an
implicit slot on the core snap from snap/implicit.go into each
particular interface definition file. This brings us closer to having
all the relevant code for a given interface be entirely defined in the
interface file itself.

The fuse-support interface had some extra complexity that was now merged
into the interface definition (it is not implicitly added on ubuntu
14.04). The interface code had to be changed from commonInterface to a
new dedicated type.

The code that adds interfaces is now moved to overlord/ifacecstate and
made private (it is only relevant there and it avoids cyclic imports).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>