tests: fix snap confine from core test to check the restart was done
This fix is to make sure the restart has been done, this is because
before there was a match in journalctl even if the permissions we not
changed. By counting the occurrencies we are sure a restart has been
done.

An initial check for the /usr/lib/snapd/snap-confine permissions is done
that will help in case of errors.