client: refactor new interface api to list + details
Currently there is one API endpoint for everything related to all the
interfaces on the system. This patch splits it into two, list and
details.

The list interface returns just the names (though this may evolve to
names + basic meta-data) and a details endpoint that returns complete
details about a single interface.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>