store: address review feedback
The neat thing about this one is probably that the new (private) store
method refreshForCandidate, called by both LookupRefresh and
ListRefresh to actually contact the store, has been made mockable. The
tests for LookupRefresh are done using this, and I'll be rewriting the
ListRefresh ones to do similarly once I'm happy with the coverage for
the helper funcs (which this branch also adds).