cmd/hotfix: add hotfix tool
The hotfix tool is intended to help people that are tracking edge and
get into funny state that is hard to recover manually (e.g. by
refreshing to another core revision) but can still be fixed
automatically.

Along with the simple tool we also add the first hotfix, the one that is
affecting people on the core revision 1968, where core refresh fails
with "snap-update-ns" errors.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>