Reducing the size of the journalctl logs for ubuntu 14.04
This is a way to reduce the size of the journalctl logs on ubuntu 14.04.

This is needed to improve the quality of the tests bacause some of them
are using journalctl to check that the actions are done.

I also will help to make more readable the logs when a test fail.