interfaces: compose the base declaration from interfaces
Whenever a new interface is created the developer has to edit a number
of files. With the recent improvements those have shrinked to, almost,
just the interface module. This patch reduces this to *exactly* just the
new interface module. Now each interface can augment the base
declaration straight from the module file.

The split change nothing in how the policy looks like but sets the stage
for composing it out of meta-data coming from individual interfaces.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>