tests: apt auto-clean mechanism
This branch adds a mechanism to make an autoclean of the packages
installed through apt.

The objective is leave the environment as clean as possible by purging
automatically all the packages installed in the test.

For that an env variable has to be defined as => APT_AUTOCLEAN: 1. The
variable has to be defined explicitely in order to avoid be applied in
tests for example when snapd is intalled.