Wait in `snap run` for snapd to finish activating
Our snapd uses the systemd notify protocol and will only finish
initialization when all security profiles have been re-generated.

This allows us to check in `snap run` if snapd is still activating
and if so to wait a little bit. This will wait for a maximum of
5s after that the old profiles will be used. Note that systemd
itself will give up after ~100 seconds so the worst case is that
snapd fails to start on boot. In this case all snaps will start
with a delay of 5 seconds in the first 100 seconds. After that
things will be "normal" (except that the old security profiles
will be used).