tests: fix for the test postrm-purge
The fix is moves the clean up of the systemd service unit config files
to the reset.

This change will address the scenarios/tests where snapd is not in a
working state in the restore step.

A new config.sh file is added which contains the configuration needed to
setup the system service units.

Error fixing:
https://travis-ci.org/snapcore/snapd/builds/240775990