systemd: refactor service ops to also be exposed more directly
This has a small knock-on effect on some clients of the systemd
package because of a rename from Stop and Restart to StopAndWait and
RestartAndWait respectively, otherwise it works in the same way. Now
exposed also are EnableNow/DisableNow for the enable-and-start and
stop-and-disable variants of enable/disable, as well as Reload, and
direct Start and Restart which don't wait for things to actually stop.