Change main store host to api.snapcraft.io
As part of the enterprise store project, the store is moving over to
having a single api.snapcraft.io host which should eventually be the
front door for all queries from devices; this will make it much easier
to put something like an edge proxy in place, because only one URL will
need to be overridden.

At the moment, api.snapcraft.io only provides the non-legacy endpoints
that used to be part of Click Package Index, but it will soon be
expanded to provide others as well.  This shifts snapd over to the new
host name.  The older endpoints will of course be maintained for as long
as is necessary.

I've dropped the trailing `v1/` from the URL, since we'll probably have
a `v2/` at some point but that may be on an endpoint-by-endpoint basis,
so it doesn't make sense to pin all of snapd's API calls to the same
version.

The preferred way to set an alternate store API URL is now
`SNAPPY_FORCE_API_URL=https://api.example.org/api/`.
`SNAPPY_USE_STAGING_STORE=1` still works as an abbreviated way to point
at staging.