many: "os" -> "core" snap type
this switches everything so it's a core snap type, not an os snap
type. Includes a client-visible change, but confirmed with
gnome-software that they don't key off of this anywhere. Also includes
backwards compatibility so that we still understand type:os, and
still serialises type:core as type:os so we don't need a flag day.