cmd,client,daemon: expose "force devmode" in sysinfo
The GNOME software center would like to show appopriate icons if a snap
application is confined and while each snap shows its confinement
information the whole system may be unable to offer effective
confinement. This patch adds a "confinement" key to the sysinfo
interface and switches the existing "snap debug confinement" command to
use it, rather than implement the logic in the client.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>