store: talk to api.snapcraft.io for purchases
api.snapcraft.io now has proxy endpoints for the purchases requests
needed by snapd, so we can use that to get us closer to consolidating
snapd's store requests onto a single device gateway.

The buy URL is now separate from orders just because that was a bit
easier to implement on the store side.