don't export snap-specific XDG_RUNTIME_DIRECTORY
Wayland compositors hard-code /run/user/<uid> for the wayland sockets which
breaks clients that honor XDG_RUNTIME_DIR when snapd sets XDG_RUNTIME_DIR to
/run/user/uid/snap.name. While clients could be made to use a symlink from
this directory to the real server socket (so the clients could find the
server), part of the wayland protocol is to setup a per-client socket, which
ends up in /run/user/uid/snap.name and the server can't find it. This is
detailed in:

  https://forum.snapcraft.io/t/wayland-dconf-and-xdg-runtime-dir/186/4

Add UserXdgRuntimeDirGlobal which will return /run/user/<uid> and adjust
snapenv.go to use this (ie, export XDG_RUNTIME_DIR=/run/user/<uid>). We still
allow via security policy use of /run/user/uid/snap.name, but apps just will
not be forced to use it any more. As such, we leave UserXdgRuntimeDir() and
friends since this is used in the cleanup code for removing these (now
optional) snap-specific directories (ie, /run/user/uid/snap.name).