cmd: fix broken double re-executiong from snapd 2.21
This patch fixes a bug where the snap command from distribution package
containing snapd 2.21 will misbehave when re-executing for the second
time in the context of a snap using classic confinement.

In 2.21 snapd used one variable SNAP_REEXEC to indicate that re-exec has
occured. Upon reexecution we would set SNAP_REEXEC=0 to say that we
don't need to reexecute anymore. Future versions of snapd split that
into SNAP_REEXEC (preference) and SNAP_DID_REEXEC (fact).

A re-executing 2.21 snapd will thus set the *preference* respected by
future version of snap that will no longer re-execute. In the case of a
classically confined snap that runs snap applications itself this will
cause the snap to misbehave as 2.26.13-installed /usr/bin/snap will no
longer put certain configuration files that 2.21 not-reexecuting
/usr/bin/snap (and thus snap-exec and snap-confine) expects.

To work around that, when we ditch the environment variable and instead
rely on /proc/self/exe symlink to tell us if we're running from the core
snap already (or not) and also if the target executable is the same one
as the one executing.

In addition we unset SNAP_REEXEC if we did re-execute and SNAP_REEXEC is
equal to 0. That value means "don't reexec" so but since we have
re-executed it must have been set by an old version of snapd and thus it
can be safely removed.

Forum: https://forum.snapcraft.io/t/snapd-2-26-9-and-conjure-up-no-longer-work/1348/
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>