Fix re-exec bug when starting from snapd 2.21
We need to special case the env handling for snapd re-execing
from 2.21. In this version of snap/snapd we did set SNAP_REEXEC=0
when we re-execed.

In this case we need to unset the reExecKey to ensure that subsequent
run of snap/snapd (e.g. when using classic confinement) will *not*
prevented from re-execing.