overlord/snapstate/backend: some copydata improvements
The state machinery needs that every handler either succeeds or fails
entirely. Currently copy-snap-data has a few half-failure modes; this
addresses that.

In particular, copying data has multiple steps that can fail
independently. This makes it so that when one of them fails, the
previous ones are cleaned up (undone), and anything that had been
trashed in the same operation is brought back from the trash.

This is easy to check, currently: simply

    snap refresh <some snap> --revision=<the current revision it's on>

and you'll see the operation fail, but also you'll notice that if the
snap had data dirs they're gone, moved to .old. This PR also enables
this operation to succeed, by explicitly checking for it in both
directions.