cmd/snap-confine: use snap-update-ns as late initializer
This patch switches snap-confine to use an *unconfined* snap-update-ns
as a late initializer of the mount namespace. By "late" I mean after the
basic structure is put in place and the .fstab processing is about to
occurr. This essentially moves all of the understanding of .fstab files
to snap-update-ns, so that it can be evolved there in one go.

In case we cannot find snap-update-ns (for whatever reason) the old
code-path that uses the same logic implemented in C is left as-is.
It will be dropped in a subsequent patch, alongside with a cleanup
of the apparmor profile.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>