snap-confine, snap-seccomp: Default to SECCOMP_RET_ERRNO
The seccomp policy has historically used SECCOMP_RET_KILL to forcefully
kill a snap process that bumps into the walls of the sandbox. However,
killing the snap is not very user friendly. Changing the policy to use
SECCOMP_RET_ERRNO to return -1 with errno set to EPERM has been desired
but the kernel would not log those denials which could leave users and
developers confused about why their applications were experiencing
errors.

The 4.14 Linux kernel contains new seccomp logging controls which allows
snapd to request SECCOMP_RET_ERRNO to be logged. This patch makes use of
the new logging controls and switches the default action of the seccomp
policy to SECCOMP_RET_ERRNO so that snaps aren't killed when the perform
an illegal system call.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>