cmd/snap-seccomp: Use seccomp's SECCOMP_RET_LOG mode for devmode
No seccomp policy was loaded for snaps installed in devmode since
seccomp didn't have the concept of "allow but log". The 4.14 Linux
kernel now has that functionality with the SECCOMP_RET_LOG action. This
patch changes the seccomp policy for devmode to log and allow any system
calls that would not be allowed if the snap was installed in jailmode.

Fixes: https://launchpad.net/bugs/1567597
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>