tests: add out-of-the-box test suite
This patch adds out-of-the-box test suite that tests how snapd behaves
out-of-the box, for regular users. There is no elaborate setup, no
package building involved. The test is juts installing snapd from the
repository (enabling a 3rd party repository in case of openSUSE) and
installs a simple snap.

The point of this test is to catch errors out in the wild, such as the
recently discovered hang (still) affecting Debian 9. For that purpose
the tests are specificall as close to what people would really do as
possible.

The suite is marked as manual and should be invoked nightly as a canary
test, to ensure we're doing okay.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>