tests: download core and ubuntu-core at most once
This patch relies on snapd's ability to resume partial downloads.  In
the one-off preparation phase we download a set of snaps, chop off the
last byte and move them to /var/lib/snapd/snaps adding the .partial
suffix to enable the download beavior. Each test that needs to pull in
one of those two fill then take advantage of the existing file and will
do minimal network traffic required to achieve this.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>