snap/snapenv: always expect /snap for $SNAP
On the "inside" of the snap execution environment we can always(1) rely on
the /snap directory to be present. On distributions where the /snap
directory on the host is redirected to /var/lib/snapd/snap this fact
should not leak into the runtime environment.

This was already done correctly but one of the tests was stale.

(1): Snaps with classic confinement don't work there anyway so that's a
non-issue.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>